pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dlib-bin>=19.24.6",
    "email-validator>=2.2.0",
    "fastapi>=0.115.12",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.5",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "starlette>=0.46.2",
    "uvicorn>=0.34.2",
]
